Update group deployer to support multiple routes
The current group deployer uses the value in the ROUTE_HOSTNAME environment
variable to specific the single hostname to map to the container group.
This commit adds support for a ADDITIONAL_HOSTNAMES environment variable
that is a comma-separated list of hostnames that should all be mapped to
the container group. All hostnames specified (the union of ROUTE_HOSTNAME
and ADDITIONAL_HOSTNAMES) will be mapped to the new group and unmapped
from any existing groups.

All hostname will share the common ROUTE_DOMAIN domain.